fix int problem

+++ aflare.py
@@ -26,7 +26,7 @@ def aflare(t, p):
     _fr = [1.00000, 1.94053, -0.175084, -2.24588, -1.12498]
     _fd = [0.689008, -1.60053, 0.302963, -0.278318]
 
-    Nflare = np.floor( (len(p)/3.0) )
+    Nflare = int( np.floor( (len(p)/3.0) ) )
 
     flare = np.zeros_like(t)
     # compute the flare model for each flare
